peers: replace ConnectionManager with Channels and ConnectionRunner
ConnectionRunner runs a bidirectional TCP connection to one peer.
It stops if the peer hasn't sent us anything during a period of 2 min.
It also sends keep alive messages at most once every 30s if no other
messages are being sent to the peer. It may live in a different
thread than channels, if it wishes so.

Channels are an inteface to be used by upper layers, and are
supposed to live in the same thread as business logic. There is one
UploadChannel and one DownloadChannel for every peer. Note that
methods taking timeout as parameter, do NOT cancel read/write
operation after the timeout, but simply return to the caller.
UploadChannel is for sending pieces to the peer. DownloadChannel
is for receiving pieces from the peer.