Reimplement channel monitors
In addition to sending and receiving messages, monitors will now
also keep track of peer connection state, i.e. local/remote choke
and local/remote interest. When a monitor has finished an async
operation, it can be inspected in order to get the latest info on
the state of the peer connection.