Add PeerStatus
PeerStatuses live in the synchronous half of the code, and their
purpose is to mirror the states of channels. The channels live in
the asynchronous part of the application and are therefore only
available one at a time, once the last async operation on the
respective channel monitor has completed.

Peer statuses can be queried at any time to get the last known state
of any channel, regardless of whether that channel/monitor is
currently performing an async operation. In order to replicate the
states more easily, the peer statuses consume
[Upload|Download]ChannelStateUpdate structs every time the respective
channel monitor becomes available.

In addition, the peer statuses hold queues of messages that should
be sent to the peer through the respective channel monitor. When
enqueuing a new message, duplicates and mutually negating messages
(e.g. a Choke followed by an Unchoke) will be ignored.

This commit also contains some cleanup in OperationController.
Changing Dispatcher Handler to return Option<_> of operations allows
to make the code slighly more readable, hopefully without performance
penalty as Option<Vec<_>> has the same size as Vec<_>.