OperationController: read torrent data jit
When sending a block to a peer, it will be filled with data
just-in-time, so that we don't keep memory allocated longer then
necessary.

In addition, when uploading data we should limit how much data we
send at once because the peer might cancel their requests. To
avoid slowing down the upload, the read timeout for upload monitor
is reduced from 1s to 100ms.