PeerStatus: upload max 10 max blocks at a time
Limit the number of blocks that can be scheduled for upload in one
go by checking the total number of bytes and capping it by
10*MAX_BLOCK_SIZE. This is because the peer might cancel some of
its requests and we should be able to check the received messages.
Since both outgoing blocks and incoming cancel messages are
received on the same monitor, it happens sequentially and we can't
let the monitor be busy for too long by scheduling a long write.

Note that in order to be able to receive the peer's cancel before
scheduling the write of the corresponding block, we have to read
multiple downloader messages at once. To achieve that, the upload
monitor will now read until no messages have been received during
the last 500ms. Since read and write are done sequentially using
the same monitor, this can potentially slow down upload by 500ms
for every 10*MAX_BLOCK_SIZE bytes.

If this will be a problem in the future, the upload channels and
monitors should be split into two parts, one for incoming and one
for outgoing messages.