Channels: prioritize sending downloader msgs over uploader msgs
When both an uploader message and a downloader message are
available to be sent to the peer, the downloader message should
be prioritized because they are smaller.

Uploader messages can be huge chunks of data that take long time
to write into the socket. If small uploader messages are available
at the same time, their repective mpsc channel might be starved
unless explicitly prioritized over the upload mpsc channel.