storage: make Storage async
Storage does blocking IO on filesystem. Now it will run in a separate
thread. Currently it doesn't spawn an extra tokio runtime, but it
is supported nevertheless, in case we'd like to run multiple
StorageRunner's on the same thread in the future.

An alternative solution would be to use `tokio::spawn_blocking()`
for every file write or read. That would allow to avoid creating a
new oneshot channel for every operation. However, that wouldn't
guarantee that the operations are performed in the same order as
they are scheduled, i.e. we could end up reading data that has been
written partially or not yet written at all.