Fix last piece bug and add seeder emulator
The last piece is often incomplete, i.e. it's shorter than the
others. The total number of bytes may therefore be lower than the
number of pieces multiplied by piece length.

Seeder is a client that emulates a seeder. It supports 1 peer and
listens for incoming pwp connections on a specified port. It claims
that it has all pieces, unchokes the peer, and handles every request
by sending zeros.